refactor(cli): move executeIf to src/
